sort views into subdirectories
