+++ packages/web/src/views/contexts/_contextId.vue
@@ -1,18 +1,18 @@
 <script setup lang="ts">
 import { useRoute } from "vue-router"
 
-import TodoList from "../components/TodoList.vue"
-import { useProjectsStore } from "@/stores/projects";
+import TodoList from "@/components/TodoList.vue"
+import { useContextsStore } from "@/stores/contexts";
 import { storeToRefs } from "pinia";
 
 const route = useRoute()
-const { projectId } = route.params
+const { contextId } = route.params
 
-const store = useProjectsStore()
+const store = useContextsStore()
 const { find } = storeToRefs(store)
 </script>
 
 <template lang="pug">
 main
-  TodoList(:project="find(projectId)")
+  TodoList(:context="find(contextId)")
 </template>
