+++ packages/web/src/views/projects/_projectId.vue
@@ -1,18 +1,18 @@
 <script setup lang="ts">
 import { useRoute } from "vue-router"
 
-import TodoList from "../components/TodoList.vue"
-import { useContextsStore } from "@/stores/contexts";
+import TodoList from "@/components/TodoList.vue"
+import { useProjectsStore } from "@/stores/projects";
 import { storeToRefs } from "pinia";
 
 const route = useRoute()
-const { contextId } = route.params
+const { projectId } = route.params
 
-const store = useContextsStore()
+const store = useProjectsStore()
 const { find } = storeToRefs(store)
 </script>
 
 <template lang="pug">
 main
-  TodoList(:context="find(contextId)")
+  TodoList(:project="find(projectId)")
 </template>
