+++ packages/web/src/views/tags/index.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import TagList from '../components/TagList.vue'
+import TagList from '@/components/TagList.vue'
 </script>
 
 <template lang="pug">
